python/requirements.txt
anyio==3.6.2
black==22.10.0
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
joblib==1.2.0
JPype1==1.4.1
konlpy==0.6.0
lxml==4.9.1
mypy-extensions==0.4.3
nltk==3.7
numpy==1.23.4
packaging==21.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
pydantic==1.10.2
pyparsing==3.0.9
regex==2022.10.31
setuptools==58.0.4
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
uvicorn==0.19.0
